Fix 400 error on GET /api/v1/products/{id}: int -> long
Product.Id is long (BIGINT) but controller and service used int.
EF Core FindAsync throws ArgumentException on type mismatch,
which ExceptionHandlingMiddleware maps to 400 Bad Request.

Route constraint changed from {id:int} to {id:long}.
All product/image service signatures updated to use long.

+++ src/VypusknykPlus.Application/Services/IImageService.cs
@@ -5,7 +5,7 @@ public interface IImageService
     /// <summary>
     /// Upload an image stream for a product. Returns the stored object key (e.g. "products/42.jpg").
     /// </summary>
-    Task<string> UploadProductImageAsync(int productId, Stream imageStream, string contentType, CancellationToken ct = default);
+    Task<string> UploadProductImageAsync(long productId, Stream imageStream, string contentType, CancellationToken ct = default);
 
     /// <summary>
     /// Delete an object by its key. Silently no-ops if the object does not exist.
